Create Final Project
Midtown is a busy commercial area and a vibrant arts hub. The High Museum of Art shows world-renowned works in a striking modern building, while Margaret Mitchell House offers tours of the former home of the “Gone With the Wind” author. Peachtree Street is a hotspot for comedy, bars and big-name shops, with eating options ranging from street food to fine dining. Large, leafy Piedmont Park offers walking trails.

Problem
I will be trying to open a new coffee Store in midtown Atlanta.
I will like to know how many Coffee stores are already in the area.
How spaced out are the stores. and how much foot traffic is in the area and what time will be the best time to be open
@hosuorji
Commit new file